refactor new creation
